Formatting and Comment Cleanup

+++ src/renderer2.hpp
@@ -18,7 +18,7 @@ struct [[clang::trivial_abi]] renderer2 : Tmp
 
    renderer2(Tmp tmp): Tmp{tmp}{};
 
-  static constinit inline uint8_t               imgIndx;
+  static constinit inline uint32_t               imgIndx;
   static constinit inline uint8_t               currentFrame;
   
   template <typename type> constexpr auto doSet( auto &s, auto f) -> std::array<type, Frames>;
